requirements.txt
python-Levenshtein==0.12.2
coredis>=3.4,<3.12
emoji>=1.6,<1.8
python-dotenv>=0.19,<0.21
sentry_sdk>=1.5,<1.7
janus>=1.0,<1.1
msgpack>=1,<1.1
aiofiles>=0.8,<0.9
httpx>=0.16,<0.24
psutil>=5.9,<5.10
decancer-py==0.1.0
git+https://github.com/Shinobou/captcha-rs
hikari@git+https://github.com/cleaner-bot/hikari.git
expirepy@git+https://github.com/cleaner-bot/expirepy.git